Move Python Tests Code into Mono-Package
* Move tests from previous packages into mono-package tests directory.
  This new directory mimics the directory structure of the new source
  directory.
* Move the Ingestion Tests over and fix problems with them.
* Move the Performance Tests over as well.
