--- performance_manager/tests/demo_events.json
@@ -1,29 +1,0 @@
-[
-    {
-        "path": "mbta-ctd-dataplatform-dev-springboard/lamp/RT_VEHICLE_POSITIONS/year=2022/month=7/day=20/hour=0/ade29e12f6144ebdaa27c86083a358b3-0.parquet"
-    },
-    {
-        "path": "mbta-ctd-dataplatform-dev-springboard/lamp/RT_VEHICLE_POSITIONS/year=2022/month=7/day=20/hour=0/eecaaf159f804e9686f4d16f816a708f-0.parquet"
-    },
-    {
-        "path": "mbta-ctd-dataplatform-dev-springboard/lamp/RT_VEHICLE_POSITIONS/year=2022/month=7/day=20/hour=1/2e14b3724450454d8bd2578e63995dae-0.parquet"
-    },
-    {
-        "path": "mbta-ctd-dataplatform-dev-springboard/lamp/RT_VEHICLE_POSITIONS/year=2022/month=7/day=20/hour=1/3338ab3b2f9b4026881f98c902f38fa9-0.parquet"
-    },
-    {
-        "path": "mbta-ctd-dataplatform-dev-springboard/lamp/RT_TRIP_UPDATES/year=2022/month=7/day=20/hour=0/1733d119787145b587c5c9f4af43a02d-0.parquet"
-    },
-    {
-        "path": "mbta-ctd-dataplatform-dev-springboard/lamp/RT_TRIP_UPDATES/year=2022/month=7/day=20/hour=0/fce9108288f04479aff70c081e3acf30-0.parquet"
-    },
-    {
-        "path": "mbta-ctd-dataplatform-dev-springboard/lamp/RT_TRIP_UPDATES/year=2022/month=7/day=20/hour=1/14e07765d4ff4a33b400953ca2da2c4d-0.parquet"
-    },
-    {
-        "path": "mbta-ctd-dataplatform-dev-springboard/lamp/RT_TRIP_UPDATES/year=2022/month=7/day=20/hour=1/d3d6db70652743159d5d24b5051d757a-0.parquet"
-    },
-    {
-        "path": "mbta-ctd-dataplatform-dev-springboard/lamp/FEED_INFO/timestamp=1661219208/6ed104fd4c724faf89ebcad5c760ab62-0.parquet"
-    }
-]
