--- performance_manager/tests/test_performance_manager.py
@@ -1,6 +1,0 @@
-from lib import __version__
-
-
-def test_version() -> None:
-    """simple test stub"""
-    assert __version__ == "0.1.0"
