--- py_gtfs_rt_ingestion/tests/test_py_gtfs_rt_ingestion.py
@@ -1,6 +1,0 @@
-from lib import __version__
-
-
-def test_version() -> None:
-    """simple test stub"""
-    assert __version__ == "0.1.0"
