+++ python_src/tests/ingestion/test_py_gtfs_rt_ingestion.py
@@ -1,0 +1,6 @@
+from lamp_py.ingestion import __version__
+
+
+def test_version() -> None:
+    """simple test stub"""
+    assert __version__ == "0.1.0"
