+++ python_src/tests/test_files/INCOMING/2022-05-08T06:04:57Z_https_cdn.mbta.com_realtime_TripUpdates_enhanced.json.gz
