+++ python_src/tests/test_resources.py
@@ -1,0 +1,6 @@
+import os
+
+test_files_dir = os.path.join(os.path.dirname(__file__), "test_files")
+
+incoming_dir = os.path.join(test_files_dir, "INCOMING")
+springboard_dir = os.path.join(test_files_dir, "SPRINGBOARD")
